av1_block_error_lp_neon: fix block_size param type
int -> intptr_t. This fixes a Control Flow Integrity (CFI) sanitizer
failure.

This also fixes a -Wmissing-prototypes warning.

This is a port of a broader change that contained the fix:
b44333201b *_neon.c: add missing rtcd includes & CONFIG check

Bug: aomedia:3416
(cherry picked from commit bfd5fa58ddf93f7a716472cddd4fdd9930bd2525)

+++ av1/encoder/arm/neon/av1_error_neon.c
@@ -12,6 +12,7 @@
 #include <assert.h>
 
 #include "config/aom_config.h"
+#include "config/av1_rtcd.h"
 
 #include "aom_dsp/aom_dsp_common.h"
 #include "aom_dsp/arm/mem_neon.h"
@@ -60,7 +61,7 @@ int64_t av1_block_error_neon(const tran_low_t *coeff, const tran_low_t *dqcoeff,
 }
 
 int64_t av1_block_error_lp_neon(const int16_t *coeff, const int16_t *dqcoeff,
-                                int block_size) {
+                                intptr_t block_size) {
   int64x2_t error = vdupq_n_s64(0);
 
   assert(block_size >= 8);
